+++ templates/CHANGELOG.md
@@ -3,6 +3,6 @@
 {% for change in changes %}
 ### {{ change.title }}
 {% for entry in change.entries %}
-- {{ entry }}
+- {{ entry.message }}
 {% endfor %}
 {% endfor %}